Further refactoring of backend/unit test classes
