+++ app/src/main/res/layout/fragment_tab2.xml
@@ -5,7 +5,7 @@
     android:id="@+id/fragment_tab2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.tabs.Tab2">
+    tools:context=".view.tabs.FavorPage">
 
     <!-- TODO: Update blank fragment layout -->
